moving date and time settings in viewModel, changing .startWith() with .isBlank()
